Fixing routing for deployment in Vercel
