+++ src/app/addrecipe/page.tsx
@@ -1,7 +1,7 @@
 import Footer from '@/components/Footer';
 import './addrecipepage.css';
 import AddRecipeForm from '@/components/AddRecipeForm';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/lib/auth';
 import { loggedInProtectedPage } from '@/lib/page-protection';
 import { getServerSession } from 'next-auth';
 import { notFound } from 'next/navigation';
