+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,96 +1,5 @@
-/* eslint-disable arrow-body-style */
-import { comparePasswords } from '@/lib/auth';
-import NextAuth, { type NextAuthOptions } from 'next-auth';
-import CredentialsProvider from 'next-auth/providers/credentials';
-import { prisma } from '@/lib/prisma';
-
-export const runtime = 'nodejs';
-
-export const authOptions: NextAuthOptions = {
-  session: {
-    strategy: 'jwt',
-  },
-  providers: [
-    CredentialsProvider({
-      name: 'Email and Password',
-      credentials: {
-        email: {
-          label: 'Email',
-          type: 'email',
-          placeholder: 'john@foo.com',
-        },
-        password: { label: 'Password', type: 'password' },
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials.password) {
-          return null;
-        }
-        const user = await prisma.user.findUnique({
-          where: {
-            email: credentials.email,
-          },
-        });
-        if (!user) {
-          return null;
-        }
-
-        const isPasswordValid = await comparePasswords(
-          credentials.password, 
-          user.password
-        );
-        
-        if (!isPasswordValid) {
-          return null;
-        }
-
-        return {
-          id: `${user.id}`,
-          email: user.email,
-          firstName: user.firstName,
-          lastName: user.lastName,
-          randomKey: user.role,
-        };
-      },
-    }),
-  ],
-  pages: {
-    signIn: '/auth/signin',
-    signOut: '/auth/signout',
-    //   error: '/auth/error',
-    //   verifyRequest: '/auth/verify-request',
-    //   newUser: '/auth/new-user'
-  },
-  callbacks: {
-    session: ({ session, token }) => {
-      return {
-        ...session,
-        user: {
-          ...session.user,
-          id: token.id,
-          randomKey: token.randomKey,
-          firstName: token.firstName,
-          lastName: token.lastName,
-          role: token.role,
-
-        },
-      };
-    },
-    jwt: ({ token, user }) => {
-      if (user) {
-        const u = user as unknown as any;
-        return {
-          ...token,
-          id: u.id,
-          randomKey: u.randomKey,
-          firstName: u.firstName,
-          lastName: u.lastName,
-          role: u.role,
-        };
-      }
-      return token;
-    },
-  },
-};
+import NextAuth from 'next-auth';
+import { authOptions } from '@/lib/auth';
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
