+++ src/lib/auth.ts
@@ -1,11 +1,83 @@
-'use server'
+import { compare } from 'bcrypt';
+import { type NextAuthOptions } from 'next-auth';
+import CredentialsProvider from 'next-auth/providers/credentials';
+import { prisma } from '@/lib/prisma';
 
-import { compare, hash } from 'bcrypt'
+export const authOptions: NextAuthOptions = {
+  session: {
+    strategy: 'jwt',
+  },
+  providers: [
+    CredentialsProvider({
+      name: 'Email and Password',
+      credentials: {
+        email: {
+          label: 'Email',
+          type: 'email',
+          placeholder: 'john@foo.com',
+        },
+        password: { label: 'Password', type: 'password' },
+      },
+      async authorize(credentials) {
+        if (!credentials?.email || !credentials.password) {
+          return null;
+        }
+        const user = await prisma.user.findUnique({
+          where: {
+            email: credentials.email,
+          },
+        });
+        if (!user) {
+          return null;
+        }
 
-export async function hashPassword(password: string) {
-  return hash(password, 10)
-}
+        const isPasswordValid = await compare(credentials.password, user.password);
+        if (!isPasswordValid) {
+          return null;
+        }
 
-export async function comparePasswords(plainPassword: string, hashedPassword: string) {
-  return compare(plainPassword, hashedPassword)
-} 
+        return {
+          id: `${user.id}`,
+          email: user.email,
+          firstName: user.firstName,
+          lastName: user.lastName,
+          randomKey: user.role,
+        };
+      },
+    }),
+  ],
+  pages: {
+    signIn: '/auth/signin',
+    signOut: '/auth/signout',
+    // error: '/auth/error',
+    // verifyRequest: '/auth/verify-request',
+    // newUser: '/auth/new-user'
+  },
+  callbacks: {
+    session: ({ session, token }) => {
+      return {
+        ...session,
+        user: {
+          ...session.user,
+          id: token.id,
+          randomKey: token.randomKey,
+          firstName: token.firstName,
+          lastName: token.lastName,
+        },
+      };
+    },
+    jwt: ({ token, user }) => {
+      if (user) {
+        const u = user as unknown as any;
+        return {
+          ...token,
+          id: u.id,
+          randomKey: u.randomKey,
+          firstName: u.firstName,
+          lastName: u.lastName,
+        };
+      }
+      return token;
+    },
+  },
+};
